+++ packages/core/package.json
@@ -53,9 +53,9 @@
     "@multiformats/base-x": "^4.0.1",
     "@noble/curves": "^2.0.1",
     "@noble/hashes": "^2.0.1",
-    "@owf/cose": "0.3.0-alpha-20260527090717",
-    "@owf/mdoc": "0.7.0-alpha-20260527101117",
-    "@owf/token-status-list": "0.3.0-alpha-20260527090717",
+    "@owf/cose": "0.3.0-alpha-20260527112128",
+    "@owf/mdoc": "0.7.0-alpha-20260527113338",
+    "@owf/token-status-list": "0.3.0-alpha-20260527112128",
     "@peculiar/asn1-ecc": "^2.6.0",
     "@peculiar/asn1-rsa": "^2.6.0",
     "@peculiar/asn1-schema": "^2.6.0",
