+++ packages/core/src/crypto/contexts/mac0Context.ts
@@ -1,13 +1,18 @@
 import { type Mac0Context } from '@owf/cose'
 import type { AgentContext } from '../../agent'
 import { CredoError } from '../../error'
-import { KeyManagementApi, type KnownJwaSignatureAlgorithm } from '../../modules/kms'
+import {
+  KeyManagementApi,
+  type KmsJwkPublicAsymmetric,
+  type KnownJwaSignatureAlgorithm,
+  knownJwaFromCoseSignatureAlgorithm,
+} from '../../modules/kms'
 
 export const getMac0Context = (agentContext: AgentContext): Mac0Context => {
   const kms = agentContext.resolve(KeyManagementApi)
 
   return {
-    mac: async (input) => {
+    authenticate: async (input) => {
       if (input.key instanceof Uint8Array) {
         throw new CredoError('For mdoc authentication with mac0 a CoseKey is required, not a Uint8Array')
       }
@@ -23,6 +28,25 @@ export const getMac0Context = (agentContext: AgentContext): Mac0Context => {
       })
 
       return signature
+    },
+    verify: async (input) => {
+      const { mac0, key } = input
+      if (key instanceof Uint8Array) {
+        throw new CredoError('For mdoc authentication verification with mac0 a CoseKey is required, not a Uint8Array')
+      }
+
+      const algorithm = knownJwaFromCoseSignatureAlgorithm(mac0.signatureAlgorithmName)
+
+      const { verified } = await kms.verify({
+        key: {
+          publicJwk: key.jwk as KmsJwkPublicAsymmetric,
+        },
+        data: mac0.toBeAuthenticated,
+        algorithm,
+        signature: mac0.tag,
+      })
+
+      return verified
     },
   } satisfies Mac0Context
 }
